% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{executeRapper}
\alias{executeRapper}
\title{Execute A "Rapped" Model}
\usage{
executeRapper(rapper)
}
\arguments{
\item{rapper}{A model built as a \code{\link{rapper}} object.}
}
\value{
A list containing the return values from the model \code{execute}
  function (see \code{\link{rapper}}). Each item in the list represents model
  output for each timestep.
}
\description{
Iterative execution of a simulation model built as a \code{\link{rapper}}
object.
}
