% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rapper}
\alias{rapper}
\title{Create Rapper Object}
\usage{
rapper(execute, drivingValues, ..., initValues = list())
}
\arguments{
\item{execute}{A function containing model code. This function will be
executed iteratively, once for each timestep (see \code{drivingValues}
argument). The function should accept one and only one argument
representing the \code{environment} within which model parameters, driving
values, and state variables exist, typically a \code{rapper} object.}

\item{drivingValues}{A \code{data.frame} with columns named for driving
variables required by the function associated with the \code{execute} argument.
Each row represents the values of the driving variables for one timestep.
The \code{row.names} must be equal to \code{0:nTimeSteps} where nTimeSteps
is the number of timesteps to execute the simulation.}

\item{...}{Named arguments describing model parameters or inital condition
for state variables where the name of the argument represents the variable
name that will store the state variable or parameter value within the rapper environment.}

\item{initValues}{An alternative way to specify parameters and initial values
of state variables using a named list. May be used in conjunction with
\code{...} arguments.}
}
\value{
Returns a rapper object (an \code{environment} with the \code{class}
  attribute set to "rapper"). The rapper object contains \code{names} bound
  to parameter values and initial state variable values passed in by the
  user. It also contains a sub-\code{environment} called ".config" containing
  the driving variables \code{data.frame} and the \code{execute} function
  passed in by the user.
}
\description{
Creates an S3 object consisting of an environment with the \code{class}
attribute set to "rapper." This object is then passed to
\code{\link{executeRapper}} to execute a model in the rapper environment.
}
